requirements.txt
altgraph==0.17.2
amqp==5.0.9
antiorm==1.2.1
APScheduler==3.8.1
billiard==3.6.4.0
cachetools==4.2.4
celery==5.2.1
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
daemon==1.2
DateTime==4.3
db==0.1.1
db-sqlite3==0.0.1
Deprecated==1.2.13
distlib==0.3.4
distro==1.6.0
docutils==0.18.1
filelock==3.4.2
Flask==2.0.2
future==0.18.2
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
kombu==5.2.2
lockfile==0.12.2
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
netaddr==0.8.0
pefile==2021.9.3
platformdirs==2.4.1
prompt-toolkit==3.0.24
protobuf==3.19.1
py2exe==0.11.0.1
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.4
pyTelegramBotAPI==4.4.0
python-crontab==2.6.0
python-daemon==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
requests==2.26.0
schedule==1.1.0
six==1.16.0
source==1.2.0
SQLAlchemy==1.4.28
timeloop==1.0.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
vine==5.0.0
virtualenv==20.13.0
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.13.3
yookassa==2.2.3
zope.interface==5.4.0
